+++ kotlin-test-junit/android/main/kotlin/com/javiersc/kotlin/test/IgnoreANDROID.kt
@@ -1,5 +1,0 @@
-package com.javiersc.kotlin.test
-
-import org.junit.Ignore
-
-public actual typealias IgnoreANDROID = Ignore