+++ kotlin-test-junit/common/main/kotlin/com/javiersc/kotlin/test/IgnoreIOS.kt
@@ -1,3 +1,0 @@
-package com.javiersc.kotlin.test
-
-@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION) public expect annotation class IgnoreIOS