+++ kotlin-test-junit5/android/main/kotlin/com/javiersc/kotlin/test/IgnoreANDROID.kt
@@ -1,5 +1,0 @@
-package com.javiersc.kotlin.test
-
-import org.junit.jupiter.api.Disabled
-
-public actual typealias IgnoreANDROID = Disabled