+++ kotlin-test-junit5/linux/main/kotlin/com/javiersc/kotlin/test/IgnoreWATCHOS.kt
@@ -1,4 +1,0 @@
-package com.javiersc.kotlin.test
-
-@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION)
-public actual annotation class IgnoreWATCHOS