+++ kotlin-test-junit5/tvosSimulatorArm64/main/kotlin/com/javiersc/kotlin/test/IgnoreTVOSSIMULATORARM64.kt
@@ -1,5 +1,0 @@
-package com.javiersc.kotlin.test
-
-import kotlin.test.Ignore
-
-public actual typealias IgnoreTVOSSIMULATORARM64 = Ignore