+++ kotlin-test-junit5/wasm/main/kotlin/com/javiersc/kotlin/test/IgnoreWASM.kt
@@ -1,5 +1,0 @@
-package com.javiersc.kotlin.test
-
-import kotlin.test.Ignore
-
-public actual typealias IgnoreWASM = Ignore