+++ kotlin-test-testng/jvm/main/kotlin/com/javiersc/kotlin/test/IgnoreJS.kt
@@ -1,3 +1,0 @@
-package com.javiersc.kotlin.test
-
-@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION) public actual annotation class IgnoreJS