+++ kotlin-test-testng/jvm/main/kotlin/com/javiersc/kotlin/test/IgnoreJVM.kt
@@ -1,5 +1,0 @@
-package com.javiersc.kotlin.test
-
-import org.testng.annotations.Ignore
-
-public actual typealias IgnoreJVM = Ignore